refactor: 타입 데이터 이동 및 수정
